add function manager and TaskExecutionHandler
