+++ cpp/include/ray/api/ray_register.h
@@ -14,171 +14,36 @@
 
 #pragma once
 
-#include <ray/api/serializer.h>
-#include "absl/utility/utility.h"
-
-#include <boost/callable_traits.hpp>
-#include <functional>
-#include <memory>
-#include <string>
-#include <unordered_map>
+#include <ray/api/function_manager.h>
 
 namespace ray {
 namespace api {
 
-enum ErrorCode {
-  OK = 0,
-  FAIL = 1,
-};
-
-struct VoidResponse {
-  int error_code;
-  std::string error_msg;
-
-  MSGPACK_DEFINE(error_code, error_msg);
-};
-
-template <typename T>
-struct Response {
-  int error_code;
-  std::string error_msg;
-  T data;
-
-  MSGPACK_DEFINE(error_code, error_msg, data);
-};
-
-template <class, class>
-struct AddType;
-
-template <class First, class... Second>
-struct AddType<First, std::tuple<Second...>> {
-  using type = std::tuple<First, Second...>;
-};
-
-template <class First, class Second>
-using AddType_t = typename AddType<First, Second>::type;
-
-namespace internal {
-
-class Router {
- public:
-  static Router &Instance() {
-    static Router instance;
-    return instance;
-  }
-
-  template <typename Function>
-  bool RegisterRemoteFunction(std::string const &name, const Function &f) {
-    /// Now it is just support free function, it will be
-    /// improved to support member function later.
-    auto pair = func_ptr_to_key_map_.emplace((uint64_t)f, name);
-    if (!pair.second) {
-      return false;
-    }
-
-    RegisterNonMemberFunc(name, f);
-    return true;
-  }
-
-  std::string Route(const char *data, std::size_t size) {
-    std::string result;
-    do {
-      try {
-        auto p = Serializer::Deserialize<std::tuple<std::string>>(data, size);
-        auto &func_name = std::get<0>(p);
-        auto it = map_invokers_.find(func_name);
-        if (it == map_invokers_.end()) {
-          result = PackReturnValue(ErrorCode::FAIL, "unknown function: " + func_name);
-          break;
-        }
-
-        result = it->second(data, size);
-      } catch (const std::exception &ex) {
-        result = PackReturnValue(ErrorCode::FAIL, ex.what());
-      }
-    } while (0);
-
-    return result;
-  }
-
- private:
-  Router() = default;
-  ~Router() = default;
-  Router(const Router &) = delete;
-  Router(Router &&) = delete;
-
-  template <typename F, size_t... I, typename Arg, typename... Args>
-  static typename std::result_of<F(Args...)>::type CallInternal(
-      const F &f, const absl::index_sequence<I...> &, std::tuple<Arg, Args...> tup) {
-    return f(std::move(std::get<I + 1>(tup))...);
-  }
-
-  template <typename F, typename Arg, typename... Args>
-  static typename std::enable_if<
-      std::is_void<typename std::result_of<F(Args...)>::type>::value>::type
-  Call(const F &f, std::string &result, std::tuple<Arg, Args...> tp) {
-    CallInternal(f, absl::make_index_sequence<sizeof...(Args)>{}, std::move(tp));
-    result = PackReturnValue(ErrorCode::OK);
-  }
-
-  template <typename F, typename Arg, typename... Args>
-  static typename std::enable_if<
-      !std::is_void<typename std::result_of<F(Args...)>::type>::value>::type
-  Call(const F &f, std::string &result, std::tuple<Arg, Args...> tp) {
-    auto r = CallInternal(f, absl::make_index_sequence<sizeof...(Args)>{}, std::move(tp));
-    result = PackReturnValue(ErrorCode::OK, "ok", r);
-  }
-
-  template <typename Function>
-  struct Invoker {
-    static inline std::string Apply(const Function &func, const char *data, size_t size) {
-      using args_tuple = AddType_t<std::string, boost::callable_traits::args_t<Function>>;
-
-      std::string result;
-      try {
-        auto tp = Serializer::Deserialize<args_tuple>(data, size);
-        Router::Call(func, result, std::move(tp));
-      } catch (msgpack::type_error &e) {
-        result = PackReturnValue(ErrorCode::FAIL,
-                                 std::string("invalid arguments: ") + e.what());
-      } catch (const std::exception &e) {
-        result = PackReturnValue(ErrorCode::FAIL,
-                                 std::string("function execute exception: ") + e.what());
-      } catch (...) {
-        result = PackReturnValue(ErrorCode::FAIL, "unknown exception");
+inline static std::string TaskExecutionHandler(const char *data, std::size_t size) {
+  std::string result;
+  do {
+    try {
+      auto p = Serializer::Deserialize<std::tuple<std::string>>(data, size);
+      auto &func_name = std::get<0>(p);
+      auto pair = internal::FunctionManager::Instance().GetFunction(func_name);
+      if (!pair.first) {
+        result = internal::PackReturnValue(internal::ErrorCode::FAIL,
+                                           "unknown function: " + func_name, 0);
+        break;
       }
 
-      return result;
+      result = (*pair.second)(data, size);
+    } catch (const std::exception &ex) {
+      result = internal::PackReturnValue(internal::ErrorCode::FAIL, ex.what());
     }
-  };
+  } while (0);
 
-  template <typename Function>
-  void RegisterNonMemberFunc(std::string const &name, Function f) {
-    this->map_invokers_[name] = {std::bind(&Invoker<Function>::Apply, std::move(f),
-                                           std::placeholders::_1, std::placeholders::_2)};
-  }
-
-  template <typename T>
-  static std::string PackReturnValue(int error_code, std::string error_msg, T result) {
-    auto buffer = Serializer::Serialize(
-        Response<T>{error_code, std::move(error_msg), std::move(result)});
-    return std::string(buffer.data(), buffer.size());
-  }
-
-  static std::string PackReturnValue(int error_code, std::string error_msg = "ok") {
-    auto buffer = Serializer::Serialize(VoidResponse{error_code, std::move(error_msg)});
-    return std::string(buffer.data(), buffer.size());
-  }
-
-  std::unordered_map<std::string, std::function<std::string(const char *, size_t)>>
-      map_invokers_;
-  std::unordered_map<uintptr_t, std::string> func_ptr_to_key_map_;
-};
-}  // namespace internal
+  return result;
+}
 
 #define RAY_REGISTER(f)                 \
   static auto ANONYMOUS_VARIABLE(var) = \
-      internal::Router::Instance().RegisterRemoteFunction(#f, f);
+      internal::FunctionManager::Instance().RegisterRemoteFunction(#f, f);
 
 #define CONCATENATE_DIRECT(s1, s2) s1##s2
 #define CONCATENATE(s1, s2) CONCATENATE_DIRECT(s1, s2)
